# ofchat/swagger/ofchat.json
{
    "swagger":"2.0",
    "info":{
        "version":"0.1-SNAPSHOT",
        "title":"project"
    },
    "host":"example.com",
    "basePath":"/",
    "schemes":[
        "http"
    ],
    "paths":{
...
        "/restapi/v1/ask/":{
            "post":{
                "consumes":[
                ],
                "produces":[
                    "application/json",
                    "application/xml"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/AssistEntity"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/AssistEntity"
                        }
                    }
                }
            }
        },
        "/restapi/v1/ask/{userId}":{
            "get":{
                "consumes":[
                ],
                "produces":[
                    "application/json",
                    "application/xml"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"userId",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/AskQueue"
                        }
                    }
                }
            },
            "put":{
                "consumes":[
                ],
                "produces":[
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"userId",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    }
                }
            },
            "post":{
                "consumes":[
                ],
                "produces":[
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"userId",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    }
                }
            },
            "delete":{
                "consumes":[
                ],
                "produces":[
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"userId",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
        "Affiliate":{
            "properties":{
                "affiliation":{
                    "type":"string",
                    "enum":[
                        "admin",
                        "member",
                        "none",
                        "outcast",
                        "owner"
                    ]
                },
                "jid":{
                    "$ref":"#/definitions/Jid"
                },
                "nick":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "role":{
                    "type":"string",
                    "enum":[
                        "moderator",
                        "none",
                        "participant",
                        "visitor"
                    ]
                }
            }
        },
...
        "AskQueue":{
            "properties":{
                "available":{
                    "type":"boolean"
                },
                "inQueue":{
                    "type":"boolean"
                },
                "queuePosition":{
                    "type":"integer"
                },
                "queueRemainingTime":{
                    "type":"integer"
                },
                "workgroupJID":{
                    "type":"string"
                }
            }
        },
        "AssistEntity":{
            "properties":{
                "agent":{
                    "type":"string"
                },
                "chatroom":{
                    "$ref":"#/definitions/MultiUserChat"
                },
                "company":{
                    "type":"string"
                },
                "country":{
                    "type":"string"
                },
                "emailAddress":{
                    "type":"string"
                },
                "groupchat":{
                    "type":"string"
                },
                "message":{
                    "type":"string"
                },
                "nickname":{
                    "type":"string"
                },
                "product":{
                    "type":"string"
                },
                "question":{
                    "type":"string"
                },
                "sender":{
                    "type":"string"
                },
                "state":{
                    "type":"string"
                },
                "url":{
                    "type":"string"
                },
                "userID":{
                    "type":"string"
                },
                "workgroup":{
                    "type":"string"
                }
            }
        },
...
        "DataForm":{
            "properties":{
                "elementName":{
                    "type":"string"
                },
                "extensionElements":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Element"
                    }
                },
                "field":{
                    "$ref":"#/definitions/FormField"
                },
                "fields":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/FormField"
                    }
                },
                "hiddenFormTypeField":{
                    "$ref":"#/definitions/FormField"
                },
                "instructions":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "items":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/DataForm$Item"
                    }
                },
                "namespace":{
                    "type":"string"
                },
                "reportedData":{
                    "$ref":"#/definitions/DataForm$ReportedData"
                },
                "title":{
                    "type":"string"
                },
                "type":{
                    "type":"string",
                    "enum":[
                        "cancel",
                        "form",
                        "result",
                        "submit"
                    ]
                }
            }
        },
        "DataForm$Item":{
            "properties":{
                "fields":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/FormField"
                    }
                }
            }
        },
        "DataForm$ReportedData":{
            "properties":{
                "fields":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/FormField"
                    }
                }
            }
        },
        "Domainpart":{
            "properties":{
            }
        },
        "Element":{
            "properties":{
            }
        },
        "EntityBareJid":{
            "properties":{
                "domain":{
                    "$ref":"#/definitions/Domainpart"
                },
                "domainBareJid":{
                    "type":"boolean"
                },
                "domainFullJid":{
                    "type":"boolean"
                },
                "entityBareJid":{
                    "type":"boolean"
                },
                "entityFullJid":{
                    "type":"boolean"
                },
                "entityJid":{
                    "type":"boolean"
                },
                "localpart":{
                    "$ref":"#/definitions/Localpart"
                },
                "localpartOrNull":{
                    "$ref":"#/definitions/Localpart"
                },
                "localpartOrThrow":{
                    "$ref":"#/definitions/Localpart"
                },
                "parentOf":{
                    "type":"boolean"
                },
                "resourceOrEmpty":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "resourceOrNull":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "resourceOrThrow":{
                    "$ref":"#/definitions/Resourcepart"
                }
            }
        },
        "EntityFullJid":{
            "properties":{
                "domain":{
                    "$ref":"#/definitions/Domainpart"
                },
                "domainBareJid":{
                    "type":"boolean"
                },
                "domainFullJid":{
                    "type":"boolean"
                },
                "entityBareJid":{
                    "type":"boolean"
                },
                "entityFullJid":{
                    "type":"boolean"
                },
                "entityJid":{
                    "type":"boolean"
                },
                "localpart":{
                    "$ref":"#/definitions/Localpart"
                },
                "localpartOrNull":{
                    "$ref":"#/definitions/Localpart"
                },
                "localpartOrThrow":{
                    "$ref":"#/definitions/Localpart"
                },
                "parentOf":{
                    "type":"boolean"
                },
                "resourceOrEmpty":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "resourceOrNull":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "resourceOrThrow":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "resourcepart":{
                    "$ref":"#/definitions/Resourcepart"
                }
            }
        },
        "ExtensionElement":{
            "properties":{
                "elementName":{
                    "type":"string"
                },
                "namespace":{
                    "type":"string"
                }
            }
        },
        "Form":{
            "properties":{
                "dataFormToSend":{
                    "$ref":"#/definitions/DataForm"
                },
                "field":{
                    "$ref":"#/definitions/FormField"
                },
                "fields":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/FormField"
                    }
                },
                "instructions":{
                    "type":"string"
                },
                "title":{
                    "type":"string"
                },
                "type":{
                    "type":"string",
                    "enum":[
                        "cancel",
                        "form",
                        "result",
                        "submit"
                    ]
                }
            }
        },
        "FormField":{
            "properties":{
                "description":{
                    "type":"string"
                },
                "elementName":{
                    "type":"string"
                },
                "label":{
                    "type":"string"
                },
                "options":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/FormField$Option"
                    }
                },
                "required":{
                    "type":"boolean"
                },
                "type":{
                    "type":"string",
                    "enum":[
                        "bool",
                        "fixed",
                        "hidden",
                        "jid_multi",
                        "jid_single",
                        "list_multi",
                        "list_single",
                        "text_multi",
                        "text_private",
                        "text_single"
                    ]
                },
                "validateElement":{
                    "$ref":"#/definitions/ValidateElement"
                },
                "values":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "variable":{
                    "type":"string"
                }
            }
        },
        "FormField$Option":{
            "properties":{
                "elementName":{
                    "type":"string"
                },
                "label":{
                    "type":"string"
                },
                "value":{
                    "type":"string"
                }
            }
        },
...
        "Jid":{
            "properties":{
                "domain":{
                    "$ref":"#/definitions/Domainpart"
                },
                "domainBareJid":{
                    "type":"boolean"
                },
                "domainFullJid":{
                    "type":"boolean"
                },
                "entityBareJid":{
                    "type":"boolean"
                },
                "entityFullJid":{
                    "type":"boolean"
                },
                "entityJid":{
                    "type":"boolean"
                },
                "localpartOrNull":{
                    "$ref":"#/definitions/Localpart"
                },
                "localpartOrThrow":{
                    "$ref":"#/definitions/Localpart"
                },
                "parentOf":{
                    "type":"boolean"
                },
                "resourceOrEmpty":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "resourceOrNull":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "resourceOrThrow":{
                    "$ref":"#/definitions/Resourcepart"
                }
            }
        },
        "Localpart":{
            "properties":{
            }
        },
...
        "MucConfigFormManager":{
            "properties":{
            }
        },
        "MucEnterConfiguration$Builder":{
            "properties":{
            }
        },
        "MultiUserChat":{
            "properties":{
                "admins":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Affiliate"
                    }
                },
                "configFormManager":{
                    "$ref":"#/definitions/MucConfigFormManager"
                },
                "configurationForm":{
                    "$ref":"#/definitions/Form"
                },
                "enterConfigurationBuilder":{
                    "$ref":"#/definitions/MucEnterConfiguration$Builder"
                },
                "joined":{
                    "type":"boolean"
                },
                "members":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Affiliate"
                    }
                },
                "moderators":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Occupant"
                    }
                },
                "nickname":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "occupant":{
                    "$ref":"#/definitions/Occupant"
                },
                "occupantPresence":{
                    "$ref":"#/definitions/Presence"
                },
                "occupants":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/EntityFullJid"
                    }
                },
                "occupantsCount":{
                    "type":"integer"
                },
                "outcasts":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Affiliate"
                    }
                },
                "owners":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Affiliate"
                    }
                },
                "participants":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Occupant"
                    }
                },
                "registrationForm":{
                    "$ref":"#/definitions/Form"
                },
                "reservedNickname":{
                    "type":"string"
                },
                "room":{
                    "$ref":"#/definitions/EntityBareJid"
                },
                "subject":{
                    "type":"string"
                }
            }
        },
        "Occupant":{
            "properties":{
                "affiliation":{
                    "type":"string",
                    "enum":[
                        "admin",
                        "member",
                        "none",
                        "outcast",
                        "owner"
                    ]
                },
                "jid":{
                    "$ref":"#/definitions/Jid"
                },
                "nick":{
                    "$ref":"#/definitions/Resourcepart"
                },
                "role":{
                    "type":"string",
                    "enum":[
                        "moderator",
                        "none",
                        "participant",
                        "visitor"
                    ]
                }
            }
        },
...
        "Presence":{
            "properties":{
                "available":{
                    "type":"boolean"
                },
                "away":{
                    "type":"boolean"
                },
                "error":{
                    "$ref":"#/definitions/XMPPError"
                },
                "extension":{
                    "type":"object"
                },
                "extensions":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/ExtensionElement"
                    }
                },
                "from":{
                    "$ref":"#/definitions/Jid"
                },
                "language":{
                    "type":"string"
                },
                "mode":{
                    "type":"string",
                    "enum":[
                        "available",
                        "away",
                        "chat",
                        "dnd",
                        "xa"
                    ]
                },
                "packetID":{
                    "type":"string"
                },
                "priority":{
                    "type":"integer"
                },
                "stanzaId":{
                    "type":"string"
                },
                "status":{
                    "type":"string"
                },
                "to":{
                    "$ref":"#/definitions/Jid"
                },
                "type":{
                    "type":"string",
                    "enum":[
                        "available",
                        "error",
                        "probe",
                        "subscribe",
                        "subscribed",
                        "unavailable",
                        "unsubscribe",
                        "unsubscribed"
                    ]
                }
            }
        },
...
        "Resourcepart":{
            "properties":{
            }
        },
...
        "Stanza":{
            "properties":{
                "error":{
                    "$ref":"#/definitions/XMPPError"
                },
                "extension":{
                    "type":"object"
                },
                "extensions":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/ExtensionElement"
                    }
                },
                "from":{
                    "$ref":"#/definitions/Jid"
                },
                "language":{
                    "type":"string"
                },
                "packetID":{
                    "type":"string"
                },
                "stanzaId":{
                    "type":"string"
                },
                "to":{
                    "$ref":"#/definitions/Jid"
                }
            }
        },
...
        "ValidateElement":{
            "properties":{
                "datatype":{
                    "type":"string"
                },
                "elementName":{
                    "type":"string"
                },
                "listRange":{
                    "$ref":"#/definitions/ValidateElement$ListRange"
                },
                "namespace":{
                    "type":"string"
                }
            }
        },
        "ValidateElement$ListRange":{
            "properties":{
                "elementName":{
                    "type":"string"
                },
                "max":{
                    "type":"integer"
                },
                "min":{
                    "type":"integer"
                }
            }
        },
...
        "XMPPError":{
            "properties":{
                "condition":{
                    "type":"string",
                    "enum":[
                        "bad_request",
                        "conflict",
                        "feature_not_implemented",
                        "forbidden",
                        "gone",
                        "internal_server_error",
                        "item_not_found",
                        "jid_malformed",
                        "not_acceptable",
                        "not_allowed",
                        "not_authorized",
                        "policy_violation",
                        "recipient_unavailable",
                        "redirect",
                        "registration_required",
                        "remote_server_not_found",
                        "remote_server_timeout",
                        "resource_constraint",
                        "service_unavailable",
                        "subscription_required",
                        "undefined_condition",
                        "unexpected_request"
                    ]
                },
                "conditionText":{
                    "type":"string"
                },
                "descriptiveText":{
                    "type":"string"
                },
                "errorGenerator":{
                    "type":"string"
                },
                "extension":{
                    "type":"object"
                },
                "stanza":{
                    "$ref":"#/definitions/Stanza"
                },
                "type":{
                    "type":"string",
                    "enum":[
                        "AUTH",
                        "CANCEL",
                        "CONTINUE",
                        "MODIFY",
                        "WAIT"
                    ]
                }
            }
        }
    }
}